requirements.txt
backports.ssl-match-hostname==3.5.0.1
cement==2.10.2
certifi==2017.7.27.1
chardet==3.0.4
dateparser==0.6.0
docker==2.5.1
docker-pycreds==0.2.1
idna==2.6
ipaddress==1.0.18
Jinja2==2.9.6
MarkupSafe==1.0
pydash==4.1.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
regex==2017.7.28
requests==2.18.4
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.33
six==1.10.0
tzlocal==1.4
urllib3==1.22
websocket-client==0.44.0
